+++ src/main/java/fr/raksrinana/fallingtree/ForgeEventSubscriber.java
@@ -6,6 +6,7 @@ import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.util.Direction;
 import net.minecraft.util.Hand;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TranslationTextComponent;
@@ -15,7 +16,6 @@ import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import javax.annotation.Nonnull;
-import java.util.Random;
 
 @Mod.EventBusSubscriber(modid = FallingTree.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE)
 public final class ForgeEventSubscriber{
@@ -50,17 +50,28 @@ public final class ForgeEventSubscriber{
 	@SubscribeEvent
 	public static void onNeighborNotifyEvent(BlockEvent.NeighborNotifyEvent event){
 		if(Config.COMMON.breakLeaves.get() && !event.getWorld().isRemote()){
-			BlockState state = event.getState();
-			Block block = state.getBlock();
-			BlockPos pos = event.getPos();
-			if(block.isIn(BlockTags.LEAVES)){
-				if(state.ticksRandomly()){
-					IWorld world = event.getWorld();
-					if(world instanceof ServerWorld){
-						block.func_225542_b_(state, (ServerWorld) world, pos, world.getRandom());
+			IWorld world = event.getWorld();
+			BlockState eventState = event.getState();
+			Block eventBlock = eventState.getBlock();
+			BlockPos eventPos = event.getPos();
+			if(eventBlock.isAir(eventState, world, eventPos)){
+				for(Direction facing : event.getNotifiedSides()){
+					BlockPos neighborPos = eventPos.offset(facing);
+					if(world.isBlockLoaded(neighborPos)){
+						BlockState neighborState = event.getWorld().getBlockState(neighborPos);
+						if(BlockTags.LEAVES.contains(neighborState.getBlock())){
+							neighborState.getBlock().func_225542_b_(neighborState, (ServerWorld) world, neighborPos, world.getRandom());
+						}
 					}
 				}
 			}
+			// if(block.isIn(BlockTags.LEAVES)){
+			// 	if(state.ticksRandomly()){
+			// 		if(world instanceof ServerWorld){
+			// 			block.func_225542_b_(state, (ServerWorld) world, pos, world.getRandom());
+			// 		}
+			// 	}
+			// }
 		}
 	}
 }